% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_hudson_k.R
\name{fst_hudson_k}
\alias{fst_hudson_k}
\title{The Generalized "Hudson" FST estimator}
\usage{
fst_hudson_k(
  X,
  labs,
  m = NA,
  ind_keep = NULL,
  loci_on_cols = FALSE,
  mem_factor = 0.7,
  mem_lim = NA
)
}
\arguments{
\item{X}{The genotype matrix (BEDMatrix, regular R matrix, or function, same as \code{popkin}).}

\item{labs}{A vector of subpopulation assignments for every individual.
At least two subpoplations must be present.}

\item{m}{The number of loci, required if \code{X} is a function (ignored otherwise).
In particular, \code{m} is obtained from \code{X} when it is a BEDMatrix or a regular R matrix.}

\item{ind_keep}{An optional vector of individuals to keep (as booleans or indexes, used to subset an R matrix).}

\item{loci_on_cols}{Determines the orientation of the genotype matrix (by default, \code{FALSE}, loci are along the rows).
Set automatically to \code{TRUE} if \code{X} is a BEDMatrix object.}

\item{mem_factor}{Proportion of available memory to use loading and processing genotypes.
Ignored if \code{mem_lim} is not \code{NA}.}

\item{mem_lim}{Memory limit in GB, used to break up genotype data into chunks for very large datasets.
Note memory usage is somewhat underestimated and is not controlled strictly.
Default in Linux and Windows is \code{mem_factor} times the free system memory, otherwise it is 1GB (OSX and other systems).}
}
\value{
A list with the following named elements, in this order:
\itemize{
\item fst: The genome-wide Fst estimate (scalar).
\item fst_loci: A vecctor of per-locus Fst estimates.
\item data: a 2-by-m matrix of statistics used in estimating Fst.
Useful to obtain a bootstrap distribution for the genome-wide Fst.
}
}
\description{
This function implements the "Hudson" FST estimator (Bhatia, Patterson, Sankararaman, and Price 2013) generalized to K subpopulations (Ochoa and Storey 2016).
Handles very large data, passed as BEDMatrix or as a regular R matrix.
Handles missing values correctly.
}
\examples{
# dimensions of simulated data
n_ind <- 100
m_loci <- 1000
k_subpops <- 10
n_data <- n_ind * m_loci

# missingness rate
miss <- 0.1

# simulate ancestral allele frequencies
# uniform (0,1)
# it'll be ok if some of these are zero
p_anc <- runif(m_loci)

# simulate some binomial data
X <- rbinom(n_data, 2, p_anc)

# sprinkle random missingness
X[ sample(X, n_data * miss) ] <- NA

# turn into a matrix
X <- matrix(X, nrow = m_loci, ncol = n_ind)

# create subpopulation labels
# k_subpops groups of equal size
labs <- ceiling( (1 : n_ind) / k_subpops )

# estimate FST using the "Hudson" formula
fst_hudson_k_obj <- fst_hudson_k(X, labs)

# the genome-wide FST estimate
fst_hudson_k_obj$fst

# vector of per-locus FST estimates
fst_hudson_k_obj$fst_loci

}
\seealso{
The popkin package.
}
