% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbr_gcta.R
\name{inbr_gcta}
\alias{inbr_gcta}
\title{2011 GCTA inbreeding estimator III}
\usage{
inbr_gcta(
  X,
  n = NA,
  mean_of_ratios = FALSE,
  loci_on_cols = FALSE,
  mem_factor = 0.7,
  mem_lim = NA,
  m_chunk_max = 1000
)
}
\arguments{
\item{X}{The genotype matrix (BEDMatrix, regular R matrix, or function, same as \code{popkin}).}

\item{n}{The number of individuals.
Required if \code{X} is a function, ignored otherwise.}

\item{mean_of_ratios}{The estimator can be computed in two broad forms.
If \code{FALSE} (default) the ratio-of-means (ROM) version is computed, which behaves more favorably and has a known asymptotic bias.
If \code{TRUE}, the mean-of-ratios (MOR) version is computed, which is more variable and has an uncharacterized bias, but is most common in the literature.}

\item{loci_on_cols}{Determines the orientation of the genotype matrix (by default, \code{FALSE}, loci are along the rows).
If \code{X} is a BEDMatrix object, the input value is ignored (set automatically to \code{TRUE} internally).}

\item{mem_factor}{Proportion of available memory to use loading and processing genotypes.
Ignored if \code{mem_lim} is not \code{NA}.}

\item{mem_lim}{Memory limit in GB, used to break up genotype data into chunks for very large datasets.
Note memory usage is somewhat underestimated and is not controlled strictly.
Default in Linux and Windows is \code{mem_factor} times the free system memory, otherwise it is 1GB (OSX and other systems).}

\item{m_chunk_max}{Sets the maximum number of loci to process at the time.
Actual number of loci loaded may be lower if memory is limiting.}
}
\value{
Inbreeding estimates
}
\description{
This function calculates the biased GCTA inbreeding estimator III described in Yang et al. (2011).
Though these estimates (MOR version) were the basis of the GRM diagonal according to that paper, the GCTA software history shows that this exact estimator was abandoned in version 0.93.0 (8 Jul 2011) in favor of \code{\link[=kinship_std]{kinship_std()}} (also MOR version), which remains in use as of writing (2022).
}
\examples{
# dimensions of simulated data
n_ind <- 100
m_loci <- 1000
n_data <- n_ind * m_loci

# missingness rate
miss <- 0.1

# simulate ancestral allele frequencies
# uniform (0,1)
# it'll be ok if some of these are zero
p_anc <- runif(m_loci)

# simulate some binomial data
X <- rbinom(n_data, 2, p_anc)

# sprinkle random missingness
X[ sample(X, n_data * miss) ] <- NA

# turn into a matrix
X <- matrix(X, nrow = m_loci, ncol = n_ind)

# estimate inbreeding
# ... ROM version (see Ochoa and Storey (2021)).
inbr_gcta_rom <- inbr_gcta(X)
# ... MOR version (from Yang et al. (2011)).
inbr_gcta_mor <- inbr_gcta(X, mean_of_ratios = TRUE)

}
\seealso{
GCTA 2011 GRM estimator ROM limit \code{\link[=kinship_gcta_limit]{kinship_gcta_limit()}}.
The limit of \code{inbr_gcta} with \code{mean_of_ratios = FALSE} is given by \code{popkin::inbr(kinship_gcta_limit(true_kinship))}.

Standard kinship estimator \code{\link[=kinship_std]{kinship_std()}} and the limit of the ROM version \code{\link[=kinship_std_limit]{kinship_std_limit()}}.

GCTA software, including history/update log.
https://yanglab.westlake.edu.cn/software/gcta/#Download
}
