% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship_std_limit.R
\name{kinship_std_limit}
\alias{kinship_std_limit}
\title{Limit of standard kinship estimator}
\usage{
kinship_std_limit(kinship, weights = NULL)
}
\arguments{
\item{kinship}{The true kinship matrix}

\item{weights}{Optional weights for individuals}
}
\value{
The biased limit of the standard kinship estimator
}
\description{
This function calculates the biased limiting expectation of the standard kinship estimator.
This limit is easily calculated given the true kinship matrix and the weights used to estimate the ancestral allele frequency.
}
\examples{
# create a dummy kinship matrix
kinship <- matrix(
    c(
        0.6, 0.1, 0,
        0.1, 0.6, 0.1,
        0, 0.1, 0.6
        ),
    nrow = 3
)
# this is its biased limit
# (uniform weights)
kinship_biased_limit <- kinship_std_limit(kinship)

}
