% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship_wg_limit.R
\name{kinship_wg_limit}
\alias{kinship_wg_limit}
\title{Limit of Weir-Goudet kinship estimator}
\usage{
kinship_wg_limit(kinship)
}
\arguments{
\item{kinship}{The true kinship matrix}
}
\value{
The biased limit of the Weir-Goudet kinship estimator
}
\description{
This function calculates the biased limiting expectation of the Weir-Goudet kinship estimator.
This limit is easily calculated given the true kinship matrix.
}
\examples{
# create a dummy kinship matrix
kinship <- matrix(
    c(
        0.6, 0.1, 0,
        0.1, 0.6, 0.1,
        0, 0.1, 0.6
        ),
    nrow = 3
)
# this is its biased limit
kinship_wg_lim <- kinship_wg_limit(kinship)

}
