% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship_gcta_limit.R
\name{kinship_gcta_limit}
\alias{kinship_gcta_limit}
\title{Limit of GCTA kinship (GRM) estimator}
\usage{
kinship_gcta_limit(kinship)
}
\arguments{
\item{kinship}{The true kinship matrix}
}
\value{
The biased limit of the GCTA kinship (GRM) estimator
}
\description{
This function calculates the biased limiting expectation of the GCTA kinship estimator, which they refer to as a GRM instead.
This limit is calculated from the true kinship matrix.
It agrees with \code{\link[=kinship_std_limit]{kinship_std_limit()}} off-diagonal, but the diagonal bias is slightly different.
}
\examples{
# create a dummy kinship matrix
kinship <- matrix(
    c(
        0.6, 0.1, 0,
        0.1, 0.6, 0.1,
        0, 0.1, 0.6
        ),
    nrow = 3
)
# this is its biased limit
kinship_gcta_lim <- kinship_gcta_limit(kinship)

}
