% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship_std.R
\name{kinship_std}
\alias{kinship_std}
\title{Standard kinship estimator}
\usage{
kinship_std(X, n = NA, mean_of_ratios = FALSE, loci_on_cols = FALSE,
  mem_factor = 0.7, mem_lim = NA)
}
\arguments{
\item{X}{The genotype matrix (BEDMatrix, regular R matrix, or function, same as \code{popkin}).}

\item{n}{The number of individuals.
Required if \code{X} is a function, ignored otherwise.}

\item{mean_of_ratios}{The standard kinship estimator can be computed in two broad forms.
If \code{FALSE} (default) the ratio-of-means version is computed, which behaves more favorably and has a known asymptotic bias.
If \code{TRUE}, the mean-of-ratios version is computed, which is more variable and has an uncharacterized bias, but is most common in the literature.}

\item{loci_on_cols}{Determines the orientation of the genotype matrix (by default, \code{FALSE}, loci are along the rows).
Set automatically to \code{TRUE} if \code{X} is a BEDMatrix object.}

\item{mem_factor}{Proportion of available memory to use loading and processing genotypes.
Ignored if \code{mem_lim} is not \code{NA}.}

\item{mem_lim}{Memory limit in GB, used to break up genotype data into chunks for very large datasets.
Note memory usage is somewhat underestimated and is not controlled strictly.
Default in Linux and Windows is \code{mem_factor} times the free system memory, otherwise it is 1GB (OSX and other systems).}
}
\value{
The estimated kinship matrix.
If \eqn{X} has names for the individuals, they will be copied to the rows and columns of this kinship matrix.
}
\description{
This function constructs the standard kinship estimates for a given genotype matrix.
Handles very large data, passed as BEDMatrix or as a regular R matrix.
Handles missing values correctly.
}
\examples{
# dimensions of simulated data
n_ind <- 100
m_loci <- 1000
n_data <- n_ind * m_loci

# missingness rate
miss <- 0.1

# simulate ancestral allele frequencies
# uniform (0,1)
# it'll be ok if some of these are zero
p_anc <- runif(m_loci)

# simulate some binomial data
X <- rbinom(n_data, 2, p_anc)

# sprinkle random missingness
X[ sample(X, n_data * miss) ] <- NA

# turn into a matrix
X <- matrix(X, nrow = m_loci, ncol = n_ind)

# estimate kinship matrices
# ... ratio-of-means version
kinship_rom <- kinship_std(X)
# ... mean-of-ratios version
kinship_mor <- kinship_std(X, mean_of_ratios = TRUE)

}
\seealso{
The popkin package.
}
